comments added
- added comments and marks
- fixed the extension folder
