+++ ExpanseTracker/Extensions/Foundation/ArrayExtension.swift
@@ -1,0 +1,19 @@
+//
+//  ArrayExtension.swift
+//  ExpanseTracker
+//
+//  Created by Rayaheen Mseri on 17/10/1446 AH.
+//
+
+import Foundation
+
+extension Array {
+    func chunked(into size: Int) -> [[Element]] {
+        stride(from: 0, to: count, by: size).map {
+            Array(self[$0..<Swift.min($0 + size, count)])
+        }
+    }
+}
+
+
+
