+++ ExpanseTracker/Extensions/SwiftUI/ViewExtensions.swift
@@ -1,5 +1,5 @@
 //
-//  Extensions.swift
+//  ViewExtensions.swift
 //  ExpenseTest1
 //
 //  Created by Tahani Ayman on 06/10/1446 AH.
