+++ ExpanseTracker/Helpers/NumberFormatterManager.swift
@@ -1,5 +1,5 @@
 //
-//  DoubleExtensions.swift
+//  NumberFormatterManager.swift
 //  ExpanseTracker
 //
 //  Created by Tahani Ayman on 12/10/1446 AH.
