+++ ExpanseTracker/Models/TimeFilter.swift
@@ -1,5 +1,5 @@
 //
-//  Tab.swift
+//  TimeFilter.swift
 //  ExpanseTracker
 //
 //  Created by Tahani Ayman on 12/10/1446 AH.
