+++ ExpanseTracker/Views/MainTabView/MainTabView.swift
@@ -20,7 +20,7 @@ struct MainTabView: View {
     var body: some View {
         ZStack {
             TabView(selection: $selectedTab) {
-                HomeView()
+                TransactionListView()
                     .tabItem {
                         Image(systemName: "house")
                         Text("Home")
