+++ ExpanseTracker/Views/Transaction/Add/AddTransactionViewModel.swift
@@ -10,50 +10,65 @@ import CoreData
 class AddTransactionViewModel: ObservableObject {
     
     private let context = PersistanceController.shared.context
-
-     func addTransaction(
-         title: String,
-         description: String,
-         amount: Double,
-         date: Date,
-         type: TransactionType,
-         selectedCategoryName: String,
-         imageData: Data?
-     ) {
-         let newTransaction = TransacionsEntity(context: context)
-         newTransaction.id = UUID().uuidString
-         newTransaction.title = title
-         newTransaction.desc = description
-         newTransaction.amount = amount
-
-         let formatter = DateFormatter()
-         formatter.dateFormat = "yyyy-MM-dd"
-         newTransaction.date = formatter.string(from: date)
-
-         newTransaction.transactionType = type.rawValue
-
-         if let imageData = imageData {
-             newTransaction.image = imageData.base64EncodedString()
-         }
-
-         let fetchRequest: NSFetchRequest<CategoryEntity> = CategoryEntity.fetchRequest()
-         fetchRequest.predicate = NSPredicate(format: "name == %@", selectedCategoryName)
-
-         do {
-             let results = try context.fetch(fetchRequest)
-             if let matchedCategory = results.first {
-                 newTransaction.category = matchedCategory
-             }
-         } catch {
-             print("Error fetching category: \(error.localizedDescription)")
-         }
-
-         do {
-             try context.save()
-             print("Transaction saved")
-         } catch {
-             print("Error saving transaction: \(error.localizedDescription)")
-         }
-     }
- }
+    
+    //MARK: - Functions
+    
+    /// Adds a new transaction to Core Data.
+    /// - Parameters:
+    ///   - title: The title of the transaction.
+    ///   - description: A short description of the transaction.
+    ///   - amount: The monetary amount of the transaction.
+    ///   - date: The date when the transaction occurred.
+    ///   - type: The transaction type (e.g., income or expense).
+    ///   - selectedCategoryName: The name of the category to associate with this transaction.
+    ///   - imageData: Optional image data to be stored with the transaction.
+    func addTransaction(
+        title: String,
+        description: String,
+        amount: Double,
+        date: Date,
+        type: TransactionType,
+        selectedCategoryName: String,
+        imageData: Data?
+    ) {
+        let newTransaction = TransacionsEntity(context: context)
+        newTransaction.id = UUID().uuidString
+        newTransaction.title = title
+        newTransaction.desc = description
+        newTransaction.amount = amount
+        
+        // Format the date to string (yyyy-MM-dd)
+        let formatter = DateFormatter()
+        formatter.dateFormat = "yyyy-MM-dd"
+        newTransaction.date = formatter.string(from: date)
+        
+        newTransaction.transactionType = type.rawValue
+        
+        // Encode image data if available
+        if let imageData = imageData {
+            newTransaction.image = imageData.base64EncodedString()
+        }
+        
+        // Fetch and Assign Category
+        let fetchRequest: NSFetchRequest<CategoryEntity> = CategoryEntity.fetchRequest()
+        fetchRequest.predicate = NSPredicate(format: "name == %@", selectedCategoryName)
+        
+        do {
+            let results = try context.fetch(fetchRequest)
+            if let matchedCategory = results.first {
+                newTransaction.category = matchedCategory
+            }
+        } catch {
+            print("Error fetching category: \(error.localizedDescription)")
+        }
+        
+        // Save to Core Data
+        do {
+            try context.save()
+            print("Transaction saved")
+        } catch {
+            print("Error saving transaction: \(error.localizedDescription)")
+        }
+    }
+}
