+++ ExpanseTracker/Views/Welcome/WelcomePage.swift
@@ -9,31 +9,29 @@
 import SwiftUI
 
 struct WelcomePage: View {
+    
     //MARK: - Variables
+    
     @EnvironmentObject var themeManager: ThemeManager
     @StateObject var auth = AuthViewModel()
-    // Get UID from user default
-       private var uid: String? {
-           UIDManager.loadUID()
-       }
+    
+    //MARK: - Auto navigation
     var body: some View {
-        //MARK: - Auto navigation
-           NavigationStack {
-               // Auto-navigation to Home if already authenticated
-               if auth.isAuthenticated {
-                   MainTabView(auth:auth)
-               } else {
-                   welcomeBody
-               }
-           }
-           .animation(.smooth, value: themeManager.isDarkMode)
-       }
-
-       
-       // keeps welcome layout separate
-
-       var welcomeBody: some View {
-           //MARK: - View
+        NavigationStack {
+            // Auto-navigation to Home if already authenticated
+            if auth.isAuthenticated {
+                MainTabView(auth:auth)
+            } else {
+                welcomeBody
+            }
+        }
+        .animation(.smooth, value: themeManager.isDarkMode)
+    }
+    
+    //MARK: - View
+    
+    // keeps welcome layout separate
+    var welcomeBody: some View {
         ZStack {
             themeManager.gradient
                 .ignoresSafeArea()
@@ -47,7 +45,7 @@ struct WelcomePage: View {
                         
                         Image("welcome")
                             .resizable()
-                            .frame(width: 350, height: 350)
+                            .frame(width: 330, height: 330)
                         
                         Text("   Welcome to SpendSmartly")
                             .font(.largeTitle)
@@ -62,6 +60,7 @@ struct WelcomePage: View {
                             .foregroundColor(themeManager.textColor.opacity(0.7))
                             .multilineTextAlignment(.center)
                             .padding(.bottom)
+                        
                         // Navigation to log in and sign up pages with the use of the custom AuthButton
                         NavigationLink(destination: LogInPage(auth:auth)) {
                             AuthButton(label: "Log in", isFilled: true)
